Save current changes it still report overflow when computing the loss value

+++ models/deepseek_coder_pipeline.py
@@ -90,7 +90,8 @@ class LlamaPipeLayer(torch.nn.Module):
         hidden_states,  labels = ipt
         attention_mask = _make_causal_mask(
             labels.shape,
-            torch.half,
+            #torch.half,
+            torch.float32,
             labels.device
         )
